Create CForkParam class for fork handling code
CForkParams is a blockheight aware class that replicates the CChainParams::Params()
method via CForkParams::ForkingParams() method which will return the proper value for
blockchain parameters which are dependent on the blockheight.

The CChainParams class is linked into sexcon-cli. Adding blockheight awareness to that class
via chainActive usage, causes sexcoin-cli to want to link CChain class which is in main.o.

Linking in main.o drags in everything so to keep things cleaner, CForkParams was created.
CForkParams is linked in with sexcoind only (not required by sexcoin-cli).